refactor: split tracker actor modules
